# docs/swagger/openapi.yaml
openapi: 3.0.1
info:
  title: 🗂️ Task Management API
  description: API zur Verwaltung von Aufgabenlisten und Tasks. Demo für Arbeitgeber
    & Portfolio.
  contact:
    name: A. El Yacoubi
    url: https://github.com/AhmedElyacoubiForJ
    email: demo@gmail.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: v1.0-demo
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/tasklists/{taskListId}/tasks/{taskId}:
    put:
      tags:
      - Tasks in TaskLists
      summary: Task innerhalb einer TaskList aktualisieren
      description: Aktualisiert einen bestehenden Task innerhalb einer TaskList.
      operationId: updateTaskInList
      parameters:
      - name: taskListId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FullUpdateTaskDto"
        required: true
      responses:
        "422":
          description: Domain-Validierung fehlgeschlagen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
        "409":
          description: Konflikt – Domainregel verletzt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
        "200":
          description: Task erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
    delete:
      tags:
      - Tasks in TaskLists
      summary: Task innerhalb einer TaskList löschen
      description: Löscht einen Task aus einer bestimmten TaskList.
      operationId: deleteTaskInList
      parameters:
      - name: taskListId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Task erfolgreich gelöscht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseVoid"
    patch:
      tags:
      - Tasks in TaskLists
      summary: Task innerhalb einer TaskList teilweise aktualisieren
      description: Aktualisiert einen bestehenden Task innerhalb einer TaskList partiell
        (PATCH).
      operationId: patchTaskInList
      parameters:
      - name: taskListId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTaskDto"
        required: true
      responses:
        "422":
          description: Domain-Validierung fehlgeschlagen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
        "200":
          description: Task erfolgreich aktualisiert (PATCH)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
        "409":
          description: Konflikt – Domainregel verletzt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskSummaryDto"
  /api/tasklists/{id}:
    get:
      tags:
      - TaskLists – CRUD
      summary: TaskList anhand der UUID abrufen
      description: "Gibt eine einzelne TaskList zurück, wenn die UUID existiert."
      operationId: getTaskList
      parameters:
      - name: id
        in: path
        description: UUID der TaskList
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: TaskList gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskListDto"
    put:
      tags:
      - TaskLists – CRUD
      summary: TaskList vollständig aktualisieren
      description: Ersetzt die vorhandene TaskList vollständig.
      operationId: updateTaskList
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskListDto"
        required: true
      responses:
        "200":
          description: TaskList erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskListDto"
    delete:
      tags:
      - TaskLists – CRUD
      summary: TaskList löschen
      description: Löscht eine bestehende TaskList.
      operationId: deleteTaskList
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Task erfolgreich gelöscht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseVoid"
...
  /api/tasklists:
    get:
      tags:
      - TaskLists – CRUD
      summary: Alle TaskLists abrufen
      description: Gibt eine Liste aller vorhandenen TaskLists zurück.
      operationId: getAllTaskLists
      responses:
        "200":
          description: Liste aller TaskLists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseListTaskListDto"
    post:
      tags:
      - TaskLists – CRUD
      summary: Neue TaskList erstellen
      description: Erstellt eine neue TaskList. Title ist erforderlich.
      operationId: createTaskList
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskListDto"
        required: true
      responses:
        "201":
          description: TaskList erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponseTaskListDto"
...
components:
  schemas:
    FullUpdateTaskDto:
      required:
      - description
      - dueDate
      - priority
      - status
      - title
      type: object
      properties:
        title:
          maxLength: 200
          minLength: 3
          required:
          - title
          type: string
          description: Titel der Aufgabe
          nullable: false
          example: Swagger konfigurieren
        description:
          maxLength: 2000
          type: string
          description: Beschreibung der Aufgabe
          nullable: true
          example: Swagger-Dokumentation für Task-Endpunkte überarbeiten
        dueDate:
          type: string
          description: Muss in der Gegenwart oder Zukunft liegen
          format: date-time
          nullable: true
        status:
          type: string
          description: Aktueller Status der Aufgabe
          example: OPEN
          enum:
          - OPEN
          - IN_PROGRESS
          - COMPLETED
        priority:
          type: string
          description: Aktualisierte Priorität der Aufgabe
          example: MEDIUM
          enum:
          - HIGH
          - MEDIUM
          - LOW
      description: DTO für ein vollständiges Update einer bestehenden Aufgabe (Full
        Update / PUT)
    APIResponseTaskSummaryDto:
      type: object
      properties:
        status:
          type: string
          description: Status der API-Antwort
          example: success
          enum:
          - SUCCESS
          - ERROR
        statusCode:
          type: integer
          description: HTTP Statuscode
          format: int32
          example: 200
        message:
          type: string
          description: Zusätzliche Nachricht zur Antwort
          example: Liste erfolgreich zurückgegeben
        errors:
          type: array
          description: Fehlerdetails bei fehlgeschlagener Anfrage
          items:
            $ref: "#/components/schemas/ApiErrorResponse"
        data:
          $ref: "#/components/schemas/TaskSummaryDto"
        timestamp:
          type: string
          description: Zeitstempel der Antwort
          format: date-time
      description: APIResponse mit TaskSummaryDto als Payload
    ApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP Statuscode
          format: int32
          example: 500
        message:
          type: string
          description: Fehlermeldung
          example: Ein unerwarteter Fehler ist aufgetreten
      description: Fehlermeldung als API-Antwort
    TaskSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: ID der Aufgabe
          format: uuid
          example: 639de273-4ad7-4367-8771-37b6f4ee030b
        title:
          type: string
          description: Titel der Aufgabe
          example: Swagger konfigurieren
        description:
          type: string
          description: Beschreibung der Aufgabe
          example: Swagger-Dokumentation für Task-Endpunkte erstellen
        dueDate:
          type: string
          description: Fälligkeitsdatum der Aufgabe
          format: date-time
        priority:
          type: string
          description: Priorität der Aufgabe
          example: HIGH
          enum:
          - HIGH
          - MEDIUM
          - LOW
        status:
          type: string
          description: Status der Aufgabe
          example: OPEN
          enum:
          - OPEN
          - IN_PROGRESS
          - COMPLETED
        taskListId:
          type: string
          description: ID der zugehörigen Task-Liste
          format: uuid
          example: 640eb884-723b-408a-869d-59b38ca44b84
      description: Zusammenfassende Informationen zu einer Aufgabe
    UpdateTaskListDto:
      type: object
      properties:
        title:
          maxLength: 100
          minLength: 3
          type: string
          description: Neuer Titel der Task-Liste
          nullable: false
          example: 🚀 Geplante Features
        description:
          maxLength: 255
          type: string
          description: Aktualisierte Beschreibung der Task-Liste
          example: Aufgaben für Sprint 2 inkl. Backend-Komponenten
        status:
          type: string
          description: Neuer Status der Task-Liste
          example: ARCHIVED
          enum:
          - ACTIVE
          - ARCHIVED
      description: DTO zur Aktualisierung einer bestehenden Task-Liste
    APIResponseTaskListDto:
      type: object
      properties:
        status:
          type: string
          description: Status der API-Antwort
          example: success
          enum:
          - SUCCESS
          - ERROR
        statusCode:
          type: integer
          description: HTTP Statuscode
          format: int32
          example: 200
        message:
          type: string
          description: Zusätzliche Nachricht zur Antwort
          example: Liste erfolgreich zurückgegeben
        errors:
          type: array
          description: Fehlerdetails bei fehlgeschlagener Anfrage
          items:
            $ref: "#/components/schemas/ApiErrorResponse"
        data:
          $ref: "#/components/schemas/TaskListDto"
        timestamp:
          type: string
          description: Zeitstempel der Antwort
          format: date-time
      description: APIResponse mit TaskListDto als Payload
    TaskListDto:
      type: object
      properties:
        id:
          type: string
          description: ID der Task-Liste
          format: uuid
          example: 640eb884-723b-408a-869d-59b38ca44b84
        title:
          type: string
          description: Titel der Task-Liste
          example: 🏁 Demo-Liste
        description:
          type: string
          description: Beschreibung der Task-Liste
          example: Beispielhafte Aufgaben für Swagger & Seeding
        created:
          type: string
          description: Erstellungszeitpunkt der Liste
          format: date-time
        updated:
          type: string
          description: Zeitpunkt der letzten Aktualisierung
          format: date-time
        count:
          type: integer
          description: Anzahl der enthaltenen Aufgaben
          format: int32
          example: 2
        progress:
          type: number
          description: Fortschritt in Prozent
          format: double
          example: 50.0
      description: Metadaten einer Task-Liste ohne Aufgaben
    CreateTaskListDto:
      type: object
      properties:
        title:
          maxLength: 100
          minLength: 3
          type: string
          description: Titel der Task-Liste
          nullable: false
          example: 🏁 Demo-Liste
        description:
          maxLength: 255
          type: string
          description: Beschreibung der Task-Liste
          example: Beispielhafte Aufgaben für Swagger & Seeding
      description: DTO zur Erstellung einer neuen Task-Liste
...
    PatchTaskDto:
      type: object
      properties:
        title:
          maxLength: 200
          minLength: 3
          type: string
          description: Neuer Titel der Aufgabe (optional)
          nullable: true
          example: Swagger konfigurieren
        description:
          maxLength: 2000
          type: string
          description: Neue Beschreibung der Aufgabe (optional)
          nullable: true
          example: Swagger-Dokumentation für Task-Endpunkte überarbeiten
        dueDate:
          type: string
          description: Muss in der Gegenwart oder Zukunft liegen
          format: date-time
          nullable: true
        status:
          type: string
          description: Neuer Status der Aufgabe (optional)
          example: IN_PROGRESS
          enum:
          - OPEN
          - IN_PROGRESS
          - COMPLETED
        priority:
          type: string
          description: Neue Priorität der Aufgabe (optional)
          example: HIGH
          enum:
          - HIGH
          - MEDIUM
          - LOW
      description: DTO für partielle Aktualisierung einer bestehenden Aufgabe (PATCH)
    APIResponseListTaskListDto:
      type: object
      properties:
        status:
          type: string
          description: Status der API-Antwort
          example: success
          enum:
          - SUCCESS
          - ERROR
        statusCode:
          type: integer
          description: HTTP Statuscode
          format: int32
          example: 200
        message:
          type: string
          description: Zusätzliche Nachricht zur Antwort
          example: Liste erfolgreich zurückgegeben
        errors:
          type: array
          description: Fehlerdetails bei fehlgeschlagener Anfrage
          items:
            $ref: "#/components/schemas/ApiErrorResponse"
        data:
          type: array
          description: Antwortdaten bei erfolgreicher Anfrage
          items:
            $ref: "#/components/schemas/TaskListDto"
        timestamp:
          type: string
          description: Zeitstempel der Antwort
          format: date-time
      description: APIResponse mit Liste von TaskListDto
...
    APIResponseVoid:
      type: object
      properties:
        status:
          type: string
          description: Status der API-Antwort
          example: success
          enum:
          - SUCCESS
          - ERROR
        statusCode:
          type: integer
          description: HTTP Statuscode
          format: int32
          example: 200
        message:
          type: string
          description: Zusätzliche Nachricht zur Antwort
          example: Liste erfolgreich zurückgegeben
        errors:
          type: array
          description: Fehlerdetails bei fehlgeschlagener Anfrage
          items:
            $ref: "#/components/schemas/ApiErrorResponse"
        data:
          type: object
          description: Antwortdaten bei erfolgreicher Anfrage
        timestamp:
          type: string
          description: Zeitstempel der Antwort
          format: date-time
      description: APIResponse ohne Payload (z. B. für DELETE)